Kafka.Consumer

Documentation

runConsumer

Arguments

:: ConsumerProperties

-> Subscription

-> (KafkaConsumer -> IO (Either KafkaError a))
A callback function to poll and handle messages

-> IO (Either KafkaError a)

Runs high-level kafka consumer.

A callback provided is expected to call pollMessage when convenient.

newConsumer :: MonadIO m => ConsumerProperties -> Subscription -> m (Either KafkaError KafkaConsumer)

Creates a kafka consumer. A new consumer MUST be closed with closeConsumer function.

assign :: MonadIO m => KafkaConsumer -> [TopicPartition] -> m KafkaError

Assigns specified partitions to a current consumer. Assigning an empty list means unassigning from all partitions that are currently assigned.

assignment :: MonadIO m => KafkaConsumer -> m (Either KafkaError (Map TopicName [PartitionId]))

Returns current consumer's assignment

subscription :: MonadIO m => KafkaConsumer -> m (Either KafkaError [(TopicName, SubscribedPartitions)])

Returns current consumer's subscription

seek :: MonadIO m => KafkaConsumer -> Timeout -> [TopicPartition] -> m (Maybe KafkaError)

pollMessage

Arguments

:: MonadIO m

=> KafkaConsumer

-> Timeout
the timeout, in milliseconds

-> m (Either KafkaError (ConsumerRecord (Maybe ByteString) (Maybe ByteString)))
Left on error or timeout, right for success

Polls the next message from a subscription

commitOffsetMessage :: MonadIO m => OffsetCommit -> KafkaConsumer -> ConsumerRecord k v -> m (Maybe KafkaError)

Commit message's offset on broker for the message's partition.

commitAllOffsets :: MonadIO m => OffsetCommit -> KafkaConsumer -> m (Maybe KafkaError)

Commit offsets for all currently assigned partitions.

commitPartitionsOffsets :: MonadIO m => OffsetCommit -> KafkaConsumer -> [TopicPartition] -> m (Maybe KafkaError)

Commit offsets for all currently assigned partitions.

closeConsumer :: MonadIO m => KafkaConsumer -> m (Maybe KafkaError)

Closes the consumer.

newtype TopicName

Topic name to be consumed

Wildcard (regex) topics are supported by the librdkafka assignor: any topic name in the topics list that is prefixed with ^ will be regex-matched to the full list of topics in the cluster and matching topics will be added to the subscription list.

Constructors

TopicName String
a simple topic name or a regex if started with ^

data OffsetCommit

Offsets commit mode

Constructors

OffsetCommit
Forces consumer to block until the broker offsets commit is done

OffsetCommitAsync
Offsets will be committed in a non-blocking way

data PartitionOffset

Constructors

PartitionOffsetBeginning

PartitionOffsetEnd

PartitionOffset Int64

PartitionOffsetStored

PartitionOffsetInvalid

data TopicPartition

Kafka topic partition structure

Constructors

TopicPartition

Fields

tpTopicName :: TopicName

tpPartition :: PartitionId

tpOffset :: PartitionOffset

data ConsumerRecord k v

Represents a received message from Kafka (i.e. used in a consumer)

Constructors

ConsumerRecord

Fields

crTopic :: !TopicName

Kafka topic this message was received from

crPartition :: !PartitionId

Kafka partition this message was received from

crOffset :: !Offset

Offset within the crPartition Kafka partition

crTimestamp :: !Timestamp

Message timestamp

crKey :: !k

crValue :: !v

data RdKafkaRespErrT

Constructors

RdKafkaRespErrBegin

RdKafkaRespErrBadMsg

RdKafkaRespErrBadCompression

RdKafkaRespErrDestroy

RdKafkaRespErrFail

RdKafkaRespErrTransport

RdKafkaRespErrCritSysResource

RdKafkaRespErrResolve

RdKafkaRespErrMsgTimedOut

RdKafkaRespErrPartitionEof

RdKafkaRespErrUnknownPartition

RdKafkaRespErrFs

RdKafkaRespErrUnknownTopic

RdKafkaRespErrAllBrokersDown

RdKafkaRespErrInvalidArg

RdKafkaRespErrTimedOut

RdKafkaRespErrQueueFull

RdKafkaRespErrIsrInsuff

RdKafkaRespErrNodeUpdate

RdKafkaRespErrSsl

RdKafkaRespErrWaitCoord

RdKafkaRespErrUnknownGroup

RdKafkaRespErrInProgress

RdKafkaRespErrPrevInProgress

RdKafkaRespErrExistingSubscription

RdKafkaRespErrAssignPartitions

RdKafkaRespErrRevokePartitions

RdKafkaRespErrConflict

RdKafkaRespErrState

RdKafkaRespErrUnknownProtocol

RdKafkaRespErrNotImplemented

RdKafkaRespErrAuthentication

RdKafkaRespErrNoOffset

RdKafkaRespErrOutdated

RdKafkaRespErrTimedOutQueue

RdKafkaRespErrUnsupportedFeature

RdKafkaRespErrWaitCache

RdKafkaRespErrIntr

RdKafkaRespErrKeySerialization

RdKafkaRespErrValueSerialization

RdKafkaRespErrKeyDeserialization

RdKafkaRespErrValueDeserialization

RdKafkaRespErrEnd

RdKafkaRespErrUnknown

RdKafkaRespErrNoError

RdKafkaRespErrOffsetOutOfRange

RdKafkaRespErrInvalidMsg

RdKafkaRespErrUnknownTopicOrPart

RdKafkaRespErrInvalidMsgSize

RdKafkaRespErrLeaderNotAvailable

RdKafkaRespErrNotLeaderForPartition

RdKafkaRespErrRequestTimedOut

RdKafkaRespErrBrokerNotAvailable

RdKafkaRespErrReplicaNotAvailable

RdKafkaRespErrMsgSizeTooLarge

RdKafkaRespErrStaleCtrlEpoch

RdKafkaRespErrOffsetMetadataTooLarge

RdKafkaRespErrNetworkException

RdKafkaRespErrGroupLoadInProgress

RdKafkaRespErrGroupCoordinatorNotAvailable

RdKafkaRespErrNotCoordinatorForGroup

RdKafkaRespErrTopicException

RdKafkaRespErrRecordListTooLarge

RdKafkaRespErrNotEnoughReplicas

RdKafkaRespErrNotEnoughReplicasAfterAppend

RdKafkaRespErrInvalidRequiredAcks

RdKafkaRespErrIllegalGeneration

RdKafkaRespErrInconsistentGroupProtocol

RdKafkaRespErrInvalidGroupId

RdKafkaRespErrUnknownMemberId

RdKafkaRespErrInvalidSessionTimeout

RdKafkaRespErrRebalanceInProgress

RdKafkaRespErrInvalidCommitOffsetSize

RdKafkaRespErrTopicAuthorizationFailed

RdKafkaRespErrGroupAuthorizationFailed

RdKafkaRespErrClusterAuthorizationFailed

RdKafkaRespErrInvalidTimestamp

RdKafkaRespErrUnsupportedSaslMechanism

RdKafkaRespErrIllegalSaslState

RdKafkaRespErrUnsupportedVersion

RdKafkaRespErrTopicAlreadyExists

RdKafkaRespErrInvalidPartitions

RdKafkaRespErrInvalidReplicationFactor

RdKafkaRespErrInvalidReplicaAssignment

RdKafkaRespErrInvalidConfig

RdKafkaRespErrNotController

RdKafkaRespErrInvalidRequest

RdKafkaRespErrUnsupportedForMessageFormat

RdKafkaRespErrPolicyViolation

RdKafkaRespErrOutOfOrderSequenceNumber

RdKafkaRespErrDuplicateSequenceNumber

RdKafkaRespErrInvalidProducerEpoch

RdKafkaRespErrInvalidTxnState

RdKafkaRespErrInvalidProducerIdMapping

RdKafkaRespErrInvalidTransactionTimeout

RdKafkaRespErrConcurrentTransactions

RdKafkaRespErrTransactionCoordinatorFenced

RdKafkaRespErrTransactionalIdAuthorizationFailed

RdKafkaRespErrSecurityDisabled

RdKafkaRespErrOperationNotAttempted

RdKafkaRespErrEndAll