Kafka.Consumer.Types

Documentation

data KafkaConsumer

Constructors

KafkaConsumer

Fields

kcKafkaPtr :: !Kafka

kcKafkaConf :: !KafkaConf

newtype ConsumerGroupId

Constructors

ConsumerGroupId String

newtype Offset

Constructors

Offset Int64

data OffsetReset

Constructors

Earliest

Latest

data SubscribedPartitions

Constructors

SubscribedPartitions [PartitionId]

SubscribedPartitionsAll

data Timestamp

Constructors

CreateTime !Millis

LogAppendTime !Millis

NoTimestamp

data OffsetCommit

Offsets commit mode

Constructors

OffsetCommit
Forces consumer to block until the broker offsets commit is done

OffsetCommitAsync
Offsets will be committed in a non-blocking way

data OffsetStoreSync

Indicates how offsets are to be synced to disk

Constructors

OffsetSyncDisable
Do not sync offsets (in Kafka: -1)

OffsetSyncImmediate
Sync immediately after each offset commit (in Kafka: 0)

OffsetSyncInterval Int
Sync after specified interval in millis

data OffsetStoreMethod

Indicates the method of storing the offsets

Constructors

OffsetStoreBroker
Offsets are stored in Kafka broker (preferred)

OffsetStoreFile FilePath OffsetStoreSync
Offsets are stored in a file (and synced to disk according to the sync policy)

data TopicPartition

Kafka topic partition structure

Constructors

TopicPartition

Fields

tpTopicName :: TopicName

tpPartition :: PartitionId

tpOffset :: PartitionOffset

data ConsumerRecord k v

Represents a received message from Kafka (i.e. used in a consumer)

Constructors

ConsumerRecord

Fields

crTopic :: !TopicName

Kafka topic this message was received from

crPartition :: !PartitionId

Kafka partition this message was received from

crOffset :: !Offset

Offset within the crPartition Kafka partition

crTimestamp :: !Timestamp

Message timestamp

crKey :: !k

crValue :: !v

crMapKey :: (k -> k') -> ConsumerRecord k v -> ConsumerRecord k' v

crMapValue :: (v -> v') -> ConsumerRecord k v -> ConsumerRecord k v'

crMapKV :: (k -> k') -> (v -> v') -> ConsumerRecord k v -> ConsumerRecord k' v'

sequenceFirst :: (Bitraversable t, Applicative f) => t (f k) v -> f (t k v)

traverseFirst :: (Bitraversable t, Applicative f) => (k -> f k') -> t k v -> f (t k' v)

traverseFirstM :: (Bitraversable t, Applicative f, Monad m) => (k -> m (f k')) -> t k v -> m (f (t k' v))

traverseM :: (Traversable t, Applicative f, Monad m) => (v -> m (f v')) -> t v -> m (f (t v'))

bitraverseM :: (Bitraversable t, Applicative f, Monad m) => (k -> m (f k')) -> (v -> m (f v')) -> t k v -> m (f (t k' v'))

data PartitionOffset

Constructors

PartitionOffsetBeginning

PartitionOffsetEnd

PartitionOffset Int64

PartitionOffsetStored

PartitionOffsetInvalid

data CommitPartitionOffset

Constructors

CommitPartitionOffset

Fields

cpoTopicName :: !TopicName

cpoPartitionId :: !PartitionId

cpoOffset :: !Offset