Kafka.Metadata

Documentation

data KafkaMetadata

Constructors

KafkaMetadata

Fields

kmBrokers :: [BrokerMetadata]

kmTopics :: [TopicMetadata]

kmOrigBroker :: !BrokerId

data BrokerMetadata

Constructors

BrokerMetadata

Fields

bmBrokerId :: !BrokerId

bmBrokerHost :: !String

bmBrokerPort :: !Int

data TopicMetadata

Constructors

TopicMetadata

Fields

tmTopicName :: !TopicName

tmPartitions :: [PartitionMetadata]

tmError :: Maybe KafkaError

data PartitionMetadata

Constructors

PartitionMetadata

Fields

pmPartitionId :: !PartitionId

pmError :: Maybe KafkaError

pmLeader :: !BrokerId

pmReplicas :: [BrokerId]

pmInSyncReplicas :: [BrokerId]

data WatermarkOffsets

Constructors

WatermarkOffsets

Fields

woTopicName :: !TopicName

woPartitionId :: !PartitionId

woLowWatermark :: !Offset

woHighWatermark :: !Offset

newtype GroupMemberId

Constructors

GroupMemberId String

data GroupMemberInfo

Constructors

GroupMemberInfo

Fields

gmiMemberId :: !GroupMemberId

gmiClientId :: !ClientId

gmiClientHost :: !String

gmiMetadata :: !ByteString

gmiAssignment :: !ByteString

newtype GroupProtocolType

Constructors

GroupProtocolType String

newtype GroupProtocol

Constructors

GroupProtocol String

data GroupState

Constructors

GroupPreparingRebalance
Group is preparing to rebalance

GroupEmpty
Group has no more members, but lingers until all offsets have expired

GroupAwaitingSync
Group is awaiting state assignment from the leader

GroupStable
Group is stable

GroupDead
Group has no more members and its metadata is being removed

data GroupInfo

Constructors

GroupInfo

Fields

giGroup :: !ConsumerGroupId

giError :: Maybe KafkaError

giState :: !GroupState

giProtocolType :: !GroupProtocolType

giProtocol :: !GroupProtocol

giMembers :: [GroupMemberInfo]

allTopicsMetadata :: (MonadIO m, HasKafka k) => k -> Timeout -> m (Either KafkaError KafkaMetadata)

Returns metadata for all topics in the cluster

topicMetadata :: (MonadIO m, HasKafka k) => k -> Timeout -> TopicName -> m (Either KafkaError KafkaMetadata)

Returns metadata only for specified topic

watermarkOffsets :: (MonadIO m, HasKafka k) => k -> Timeout -> TopicName -> m [Either KafkaError WatermarkOffsets]

Query broker for low (oldestbeginning) and high (newestend) offsets for a given topic.

watermarkOffsets' :: (MonadIO m, HasKafka k) => k -> Timeout -> TopicMetadata -> m [Either KafkaError WatermarkOffsets]

Query broker for low (oldestbeginning) and high (newestend) offsets for a given topic.

partitionWatermarkOffsets :: (MonadIO m, HasKafka k) => k -> Timeout -> TopicName -> PartitionId -> m (Either KafkaError WatermarkOffsets)

Query broker for low (oldestbeginning) and high (newestend) offsets for a specific partition

offsetsForTime :: (MonadIO m, HasKafka k) => k -> Timeout -> Millis -> [(TopicName, PartitionId)] -> m (Either KafkaError [TopicPartition])

Look up the offsets for the given partitions by timestamp.

The returned offset for each partition is the earliest offset whose timestamp is greater than or equal to the given timestamp in the corresponding partition.

offsetsForTime' :: (MonadIO m, HasKafka k) => k -> Timeout -> Millis -> TopicMetadata -> m (Either KafkaError [TopicPartition])

Look up the offsets for the given metadata by timestamp.

The returned offset for each partition is the earliest offset whose timestamp is greater than or equal to the given timestamp in the corresponding partition.

topicOffsetsForTime :: (MonadIO m, HasKafka k) => k -> Timeout -> Millis -> TopicName -> m (Either KafkaError [TopicPartition])

Look up the offsets for the given topic by timestamp.

The returned offset for each partition is the earliest offset whose timestamp is greater than or equal to the given timestamp in the corresponding partition.

allConsumerGroupsInfo :: (MonadIO m, HasKafka k) => k -> Timeout -> m (Either KafkaError [GroupInfo])

List and describe all consumer groups in cluster.

consumerGroupInfo :: (MonadIO m, HasKafka k) => k -> Timeout -> ConsumerGroupId -> m (Either KafkaError [GroupInfo])

Describe a given consumer group.