Kafka.Producer

Documentation

runProducer :: ProducerProperties -> (KafkaProducer -> IO (Either KafkaError a)) -> IO (Either KafkaError a)

Runs Kafka Producer. The callback provided is expected to call produceMessage or/and produceMessageBatch to send messages to Kafka.

newProducer :: MonadIO m => ProducerProperties -> m (Either KafkaError KafkaProducer)

Creates a new kafka producer A newly created producer must be closed with closeProducer function.

produceMessage :: MonadIO m => KafkaProducer -> ProducerRecord -> m (Maybe KafkaError)

Sends a single message. Since librdkafka is backed by a queue, this function can return before messages are sent. See flushProducer to wait for queue to empty.

produceMessageBatch

Arguments

:: MonadIO m

=> KafkaProducer

-> [ProducerRecord]

-> m [(ProducerRecord, KafkaError)]
An empty list when the operation is successful, otherwise a list of "failed" messages with corresponsing errors.

Sends a batch of messages. Returns a list of messages which it was unable to send with corresponding errors. Since librdkafka is backed by a queue, this function can return before messages are sent. See flushProducer to wait for queue to empty.

flushProducer :: MonadIO m => KafkaProducer -> m ()

Drains the outbound queue for a producer. This function is also called automatically when the producer is closed with closeProducer to ensure that all queued messages make it to Kafka.

closeProducer :: MonadIO m => KafkaProducer -> m ()

Closes the producer. Will wait until the outbound queue is drained before returning the control.

data RdKafkaRespErrT

Constructors

RdKafkaRespErrBegin

RdKafkaRespErrBadMsg

RdKafkaRespErrBadCompression

RdKafkaRespErrDestroy

RdKafkaRespErrFail

RdKafkaRespErrTransport

RdKafkaRespErrCritSysResource

RdKafkaRespErrResolve

RdKafkaRespErrMsgTimedOut

RdKafkaRespErrPartitionEof

RdKafkaRespErrUnknownPartition

RdKafkaRespErrFs

RdKafkaRespErrUnknownTopic

RdKafkaRespErrAllBrokersDown

RdKafkaRespErrInvalidArg

RdKafkaRespErrTimedOut

RdKafkaRespErrQueueFull

RdKafkaRespErrIsrInsuff

RdKafkaRespErrNodeUpdate

RdKafkaRespErrSsl

RdKafkaRespErrWaitCoord

RdKafkaRespErrUnknownGroup

RdKafkaRespErrInProgress

RdKafkaRespErrPrevInProgress

RdKafkaRespErrExistingSubscription

RdKafkaRespErrAssignPartitions

RdKafkaRespErrRevokePartitions

RdKafkaRespErrConflict

RdKafkaRespErrState

RdKafkaRespErrUnknownProtocol

RdKafkaRespErrNotImplemented

RdKafkaRespErrAuthentication

RdKafkaRespErrNoOffset

RdKafkaRespErrOutdated

RdKafkaRespErrTimedOutQueue

RdKafkaRespErrUnsupportedFeature

RdKafkaRespErrWaitCache

RdKafkaRespErrIntr

RdKafkaRespErrKeySerialization

RdKafkaRespErrValueSerialization

RdKafkaRespErrKeyDeserialization

RdKafkaRespErrValueDeserialization

RdKafkaRespErrEnd

RdKafkaRespErrUnknown

RdKafkaRespErrNoError

RdKafkaRespErrOffsetOutOfRange

RdKafkaRespErrInvalidMsg

RdKafkaRespErrUnknownTopicOrPart

RdKafkaRespErrInvalidMsgSize

RdKafkaRespErrLeaderNotAvailable

RdKafkaRespErrNotLeaderForPartition

RdKafkaRespErrRequestTimedOut

RdKafkaRespErrBrokerNotAvailable

RdKafkaRespErrReplicaNotAvailable

RdKafkaRespErrMsgSizeTooLarge

RdKafkaRespErrStaleCtrlEpoch

RdKafkaRespErrOffsetMetadataTooLarge

RdKafkaRespErrNetworkException

RdKafkaRespErrGroupLoadInProgress

RdKafkaRespErrGroupCoordinatorNotAvailable

RdKafkaRespErrNotCoordinatorForGroup

RdKafkaRespErrTopicException

RdKafkaRespErrRecordListTooLarge

RdKafkaRespErrNotEnoughReplicas

RdKafkaRespErrNotEnoughReplicasAfterAppend

RdKafkaRespErrInvalidRequiredAcks

RdKafkaRespErrIllegalGeneration

RdKafkaRespErrInconsistentGroupProtocol

RdKafkaRespErrInvalidGroupId

RdKafkaRespErrUnknownMemberId

RdKafkaRespErrInvalidSessionTimeout

RdKafkaRespErrRebalanceInProgress

RdKafkaRespErrInvalidCommitOffsetSize

RdKafkaRespErrTopicAuthorizationFailed

RdKafkaRespErrGroupAuthorizationFailed

RdKafkaRespErrClusterAuthorizationFailed

RdKafkaRespErrInvalidTimestamp

RdKafkaRespErrUnsupportedSaslMechanism

RdKafkaRespErrIllegalSaslState

RdKafkaRespErrUnsupportedVersion

RdKafkaRespErrTopicAlreadyExists

RdKafkaRespErrInvalidPartitions

RdKafkaRespErrInvalidReplicationFactor

RdKafkaRespErrInvalidReplicaAssignment

RdKafkaRespErrInvalidConfig

RdKafkaRespErrNotController

RdKafkaRespErrInvalidRequest

RdKafkaRespErrUnsupportedForMessageFormat

RdKafkaRespErrPolicyViolation

RdKafkaRespErrOutOfOrderSequenceNumber

RdKafkaRespErrDuplicateSequenceNumber

RdKafkaRespErrInvalidProducerEpoch

RdKafkaRespErrInvalidTxnState

RdKafkaRespErrInvalidProducerIdMapping

RdKafkaRespErrInvalidTransactionTimeout

RdKafkaRespErrConcurrentTransactions

RdKafkaRespErrTransactionCoordinatorFenced

RdKafkaRespErrTransactionalIdAuthorizationFailed

RdKafkaRespErrSecurityDisabled

RdKafkaRespErrOperationNotAttempted

RdKafkaRespErrEndAll