Kafka.Producer.ProducerProperties

Documentation

data ProducerProperties

Properties to create KafkaProducer.

Constructors

ProducerProperties

Fields

ppKafkaProps :: Map String String

ppTopicProps :: Map String String

ppLogLevel :: Maybe KafkaLogLevel

callbacks :: [KafkaConf -> IO ()]

brokersList :: [BrokerAddress] -> ProducerProperties

setCallback :: (KafkaConf -> IO ()) -> ProducerProperties

logLevel :: KafkaLogLevel -> ProducerProperties

Sets the logging level. Usually is used with debugOptions to configure which logs are needed.

compression :: KafkaCompressionCodec -> ProducerProperties

topicCompression :: KafkaCompressionCodec -> ProducerProperties

extraProps :: Map String String -> ProducerProperties

Any configuration options that are supported by librdkafka. The full list can be found here

suppressDisconnectLogs :: ProducerProperties

Suppresses producer disconnects logs.

It might be useful to turn this off when interacting with brokers with an aggressive connection.max.idle.ms value.

extraTopicProps :: Map String String -> ProducerProperties

Any configuration options that are supported by librdkafka. The full list can be found here

debugOptions :: [KafkaDebug] -> ProducerProperties

Sets debug features for the producer Usually is used with logLevel.

errorCallback :: HasKafkaConf k => (KafkaError -> String -> IO ()) -> k -> IO ()

logCallback :: HasKafkaConf k => (Int -> String -> String -> IO ()) -> k -> IO ()