Kafka.Producer.Types

Documentation

data KafkaProducer

Main pointer to Kafka object, which contains our brokers

Constructors

KafkaProducer

Fields

kpKafkaPtr :: !Kafka

kpKafkaConf :: !KafkaConf

kpTopicConf :: !TopicConf

data ProducerRecord

Represents messages to be enqueued onto a Kafka broker (i.e. used for a producer)

Constructors

ProducerRecord

Fields

prTopic :: !TopicName

prPartition :: !ProducePartition

prKey :: Maybe ByteString

prValue :: Maybe ByteString

data ProducePartition

Constructors

SpecifiedPartition !Int

UnassignedPartition