Kafka.Types

Documentation

class HasKafka a where

Minimal complete definition

getKafka

Methods

getKafka :: a -> Kafka

class HasKafkaConf a where

Minimal complete definition

getKafkaConf

Methods

getKafkaConf :: a -> KafkaConf

newtype BrokerId

Constructors

BrokerId Int

newtype Kafka

Constructors

Kafka RdKafkaTPtr

newtype KafkaConf

Constructors

KafkaConf RdKafkaConfTPtr

newtype TopicConf

Constructors

TopicConf RdKafkaTopicConfTPtr

newtype PartitionId

Constructors

PartitionId Int

newtype Millis

Constructors

Millis Int64

newtype ClientId

Constructors

ClientId String

newtype TopicName

Topic name to be consumed

Wildcard (regex) topics are supported by the librdkafka assignor: any topic name in the topics list that is prefixed with ^ will be regex-matched to the full list of topics in the cluster and matching topics will be added to the subscription list.

Constructors

TopicName String
a simple topic name or a regex if started with ^

newtype BrokerAddress

Kafka broker address string (e.g. broker1:9092)

Constructors

BrokerAddress String

newtype Timeout

Timeout in milliseconds

Constructors

Timeout Int

data KafkaLogLevel

Log levels for librdkafka.

Constructors

KafkaLogEmerg

KafkaLogAlert

KafkaLogCrit

KafkaLogErr

KafkaLogWarning

KafkaLogNotice

KafkaLogInfo

KafkaLogDebug

data KafkaError

Any Kafka errors

Constructors

KafkaError String

KafkaInvalidReturnValue

KafkaBadSpecification String

KafkaResponseError RdKafkaRespErrT

KafkaInvalidConfigurationValue String

KafkaUnknownConfigurationKey String

KakfaBadConfiguration

data KafkaDebug

Constructors

DebugGeneric

DebugBroker

DebugTopic

DebugMetadata

DebugQueue

DebugMsg

DebugProtocol

DebugCgrp

DebugSecurity

DebugFetch

DebugFeature

DebugAll

kafkaDebugToString :: KafkaDebug -> String

data KafkaCompressionCodec

Constructors

NoCompression

Gzip

Snappy

Lz4

kafkaCompressionCodecToString :: KafkaCompressionCodec -> String